implement shortcuts for "(today|week-start|week-end)([+-]\d+)?" patterns
