--- src/tempo_worklog_creator/util/date_util.py
@@ -1,10 +1,0 @@
-from datetime import date
-
-from tempo_worklog_creator.util.io_util import converter
-
-TODAY = "today"
-
-def process_date_input(date_str: str) -> date:
-    if date_str == TODAY:
-        return date.today()
-    return converter.structure(date_str, date)
