+++ src/tempo_worklog_creator/worklog_creator.py
@@ -18,7 +18,8 @@ from tempo_worklog_creator.constants import (
     TEMPO_WORKLOG_ID,
     ISSUE_ID,
 )
-from tempo_worklog_creator.util.io_util import load_yaml, converter
+from tempo_worklog_creator.util.io_util import load_yaml
+from tempo_worklog_creator.util.serialization import converter
 from tempo_worklog_creator.time_span import TimeSpan, FULL_DAY, MORNING, AFTERNOON
 from tempo_worklog_creator.work_log import WorkLog, WorkLogSequence
